Fix: rename graph files and fix imports
