+++ composeApp/src/commonMain/kotlin/viewmodel/GraphViewModel.kt
@@ -4,10 +4,10 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
-import lib.graph.Graph.Graph
-import lib.graph.Graph.UndirectedGraph
+import lib.graph.BaseGraph
+import lib.graph.Graph
 
-class GraphViewModel(name : String, graph : Graph = UndirectedGraph()): ViewModel() {
+class GraphViewModel(name : String, graph : BaseGraph = Graph()): ViewModel() {
     val name by mutableStateOf(name)
     val vertices = mutableStateListOf<VertexViewModel>()
     init {
