+++ lib/src/commonMain/kotlin/graph/Digraph.kt
@@ -1,6 +1,6 @@
 package lib.graph
 
-class Digraph: GraphAbstract(){
+class Digraph: BaseGraph(){
     override fun addEdge(from: Int, to: Int) {
         graph[from]?.add(to) ?: {graph[from] = mutableListOf(to)}
     }
