+++ lib/src/commonMain/kotlin/graph/Graph.kt
@@ -1,6 +1,6 @@
 package lib.graph
 
-class Graph: GraphAbstract(){
+class Graph: BaseGraph(){
     override fun addEdge(from: Int, to: Int) {
         graph[from]?.add(to) ?: {graph[from] = mutableListOf(to)}
         graph[to]?.add(from) ?: {graph[to] = mutableListOf(from)}
