+++ lib/src/commonMain/kotlin/graph/WeightedGraph.kt
@@ -1,6 +1,6 @@
 package graph
 
-import lib.graph.GraphAbstract
+import lib.graph.BaseGraph
 
 //class WeightedGraph: GraphAbstract {
 //}
